requirements.txt
PyMuPDF
numpy
sentence-transformers
faiss-cpu
chromadb
openai
